package-local-layout not promoted to layouts
